[refactor] change return type and rename file
